fix: EVE ID字段 int32→int64 溢出修复
sde_stations.station_id、eve_characters.character_id/corporation_id/alliance_id、
user_trades.station_id 从 Integer 改为 BigInteger，修复 station_id 超出 int32 范围导致交易机会加载失败。

+++ backend/app/models/sde.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Integer, Float, Boolean, ForeignKey
+from sqlalchemy import String, Integer, BigInteger, Float, Boolean, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 from app.database import Base
 
@@ -26,7 +26,7 @@ class SdeSystem(Base):
 
 class SdeStation(Base):
     __tablename__ = "sde_stations"
-    station_id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    station_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     name: Mapped[str] = mapped_column(String(200))
     name_zh: Mapped[str] = mapped_column(String(200), nullable=True)
     system_id: Mapped[int] = mapped_column(Integer, index=True)
